requirements.txt
torch==1.9.0
transformers==4.8.2
datasets==1.9.0
accelerate==0.3.0
faiss-cpu==1.7.1
faiss-gpu==1.7.1
parlai==1.3.0
ray==1.5.1
h5py==3.2.1
pyserini
python-Levenshtein
thefuzz
pytrec_eval
